requirements.txt
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.5
gunicorn==21.2.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Babel==4.0.0